Modified set_origin script

+++ src/aruco_gridboard/scripts/set_origin.py
@@ -7,24 +7,29 @@
 ##
 
 import rospy
-from pymavlink.dialects.v10 import ardupilotmega as MAV_APM
 from mavros.mavlink import convert_to_rosmsg
 from mavros_msgs.msg import Mavlink
+from pymavlink.dialects.v10 import ardupilotmega as MAV_APM
 
 # Global position of the origin
-lat = 42.56335 * 1e7   # Terni
-lon = 12.64329 * 1e7   # Terni
-alt = 163 * 1e3
+lat = int(42.56335 * 1e7)  # Terni
+lon = int(12.64329 * 1e7)  # Terni
+alt = int(163 * 1e3)
+
 
 class fifo(object):
     """ A simple buffer """
+
     def __init__(self):
         self.buf = []
+
     def write(self, data):
         self.buf += data
         return len(data)
+
     def read(self):
         return self.buf.pop(0)
+
 
 def send_message(msg, mav, pub):
     """
@@ -35,6 +40,7 @@ def send_message(msg, mav, pub):
     pub.publish(rosmsg)
 
     print("sent message %s" % msg)
+
 
 def set_global_origin(mav, pub):
     """
@@ -48,12 +54,10 @@ def set_global_origin(mav, pub):
     altitude = alt
 
     msg = MAV_APM.MAVLink_set_gps_global_origin_message(
-            target_system,
-            lattitude, 
-            longitude,
-            altitude)
+        target_system, lattitude, longitude, altitude)
 
     send_message(msg, mav, pub)
+
 
 def set_home_position(mav, pub):
     """
@@ -66,32 +70,25 @@ def set_home_position(mav, pub):
     lattitude = lat
     longitude = lon
     altitude = alt
-    
+
     x = 0
     y = 0
     z = 0
-    q = [1, 0, 0, 0]   # w x y z
+    q = [1, 0, 0, 0]  # w x y z
 
     approach_x = 0
     approach_y = 0
     approach_z = 1
 
-    msg = MAV_APM.MAVLink_set_home_position_message(
-            target_system,
-            lattitude,
-            longitude,
-            altitude,
-            x,
-            y,
-            z,
-            q,
-            approach_x,
-            approach_y,
-            approach_z)
+    msg = MAV_APM.MAVLink_set_home_position_message(target_system, lattitude,
+                                                    longitude, altitude, x, y,
+                                                    z, q, approach_x,
+                                                    approach_y, approach_z)
 
     send_message(msg, mav, pub)
 
-if __name__=="__main__":
+
+if __name__ == "__main__":
     try:
         rospy.init_node("origin_publisher")
         mavlink_pub = rospy.Publisher("/mavlink/to", Mavlink, queue_size=20)
@@ -103,7 +100,7 @@ if __name__=="__main__":
         # wait to initialize
         while mavlink_pub.get_num_connections() <= 0:
             pass
-   
+
         for _ in range(2):
             rospy.sleep(1)
             set_global_origin(mav, mavlink_pub)
